feat: [WPN-23] Publish the NPM package in github actions using semantic-release (#41)
* feat: [WPN-23] Publish the NPM package in github actions using semantic-release

* feat: change release branch name
