+++ .github/workflows/release.yml
@@ -1,0 +1,89 @@
+name: Package Release
+
+on:
+  push:
+    branches:
+      - latest-release
+      - beta
+
+permissions:
+  contents: write
+  issues: write
+  pull-requests: write
+
+jobs:
+  detect-changes:
+    runs-on: ubuntu-latest
+    outputs:
+      builder_changed: ${{ steps.filter.outputs.builder }}
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0 # Needed for changeset detection
+
+      - uses: dorny/paths-filter@v2
+        id: filter
+        with:
+          filters: |
+            builder:
+              - 'packages/builder/**'
+              # Add more packages here as repo grows
+              # package2:
+              #   - 'packages/package2/**'
+
+  release-builder:
+    needs: detect-changes
+    if: needs.detect-changes.outputs.builder_changed == 'true'
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: packages/builder
+    steps:
+      - uses: actions/checkout@v4
+        with:
+          fetch-depth: 0 # Important for semantic release to work correctly
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: "22"
+          registry-url: "https://registry.npmjs.org"
+
+      - name: Setup pnpm
+        uses: pnpm/action-setup@v2
+        with:
+          version: "10.8.1"
+          run_install: false
+
+      - name: Get pnpm store directory
+        id: pnpm-cache
+        shell: bash
+        run: |
+          echo "STORE_PATH=$(pnpm store path)" >> $GITHUB_OUTPUT
+
+      - uses: actions/cache@v3
+        name: Setup pnpm cache
+        with:
+          path: ${{ steps.pnpm-cache.outputs.STORE_PATH }}
+          key: ${{ runner.os }}-pnpm-store-${{ hashFiles('**/pnpm-lock.yaml') }}
+          restore-keys: |
+            ${{ runner.os }}-pnpm-store-
+
+      - name: Install dependencies
+        run: pnpm install --frozen-lockfile
+
+      - name: Build package
+        run: pnpm build
+
+      - name: Release
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          NODE_AUTH_TOKEN: ${{ secrets.NPM_TOKEN }}
+          NPM_TOKEN: ${{ secrets.NPM_TOKEN }}
+        run: pnpm semantic-release
+
+  # Add more jobs for additional packages as repo grows
+  # release-package2:
+  #   needs: detect-changes
+  #   if: needs.detect-changes.outputs.package2_changed == 'true'
+  #   ... (similar setup to release-builder)
